requirements.txt
torch
transformers
sentence-transformers
datasets
chromadb
pysqlite3-binary
Pillow
numpy
accelerate
